backend/functions/transcriptYoutube/requirements.txt
appwrite
pytube
deepgram-sdk
asyncio